refactor: enum 클래스들 패키지 변경
- Bridge: bridge/enums -> bridge/model/domain
- BridgeSign: bridge/enums -> bridge/model/domain
- SuccessOrFail: bridge/enums -> bridge/model/domain
